+++ podcasts/MultiSelectActionController.swift
@@ -29,18 +29,25 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
         didSet {
             selectedCountLabel.style = .primaryText02
             selectedCountLabel.themeOverride = themeOverride
+            selectedCountLabel.font = UIFont.font(ofSize: 13, weight: .bold, scalingWith: .largeTitle)
+            selectedCountLabel.adjustsFontForContentSizeCategory = true
         }
     }
 
     @IBOutlet var editButton: UIButton! {
         didSet {
             editButton.setTitle(L10n.edit, for: .normal)
+            editButton.titleLabel?.font = UIFont.font(ofSize: 15, weight: .bold, scalingWith: .largeTitle)
+            editButton.titleLabel?.adjustsFontForContentSizeCategory = true
+
         }
     }
 
     @IBOutlet var doneButton: UIButton! {
         didSet {
             doneButton.setTitle(L10n.done, for: .normal)
+            doneButton.titleLabel?.font = UIFont.font(ofSize: 15, weight: .bold, scalingWith: .largeTitle)
+            doneButton.titleLabel?.adjustsFontForContentSizeCategory = true
         }
     }
 
@@ -48,6 +55,9 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
         didSet {
             rearrangeLabel.themeOverride = themeOverride
             rearrangeLabel.text = L10n.playerActionsRearrangeTitle.localizedCapitalized
+            rearrangeLabel.numberOfLines =  2
+            rearrangeLabel.font = UIFont.font(ofSize: 18, weight: .semibold, scalingWith: .largeTitle)
+            rearrangeLabel.adjustsFontForContentSizeCategory = true
         }
     }
 
@@ -89,11 +99,10 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
     override func viewDidLoad() {
         super.viewDidLoad()
         selectedCountLabel.text = numSelectedEpisodes == 1 ? L10n.multiSelectSelectedCountSingular : L10n.multiSelectSelectedCountPlural(numSelectedEpisodes.localized())
-        actionsTable.isScrollEnabled = false
+        actionsTable.isScrollEnabled = true
         rearrangeLabel.isHidden = true
         doneButton.isHidden = true
         updateColors()
-        enableScrollingIfViewTooSmall()
 
         Analytics.track(.multiSelectViewOverflowMenuShown, properties: ["source": actionDelegate.multiSelectViewSource])
     }
@@ -127,6 +136,10 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
     }
 
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        UITableView.automaticDimension
+    }
+
+    func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
         MultiSelectActionController.rowHeight
     }
 
@@ -186,6 +199,13 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
 
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
         if tableView.isEditing {
+            return UITableView.automaticDimension
+        }
+        return CGFloat.leastNonzeroMagnitude
+    }
+
+    func tableView(_ tableView: UITableView, estimatedHeightForHeaderInSection section: Int) -> CGFloat {
+        if tableView.isEditing {
             return 40
         }
         return CGFloat.leastNonzeroMagnitude
@@ -231,7 +251,6 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
 
         if let sheetController = sheetPresentationController {
             sheetController.animateChanges { sheetController.detents = [.large()] }
-            enableScrollingIfViewTooSmall()
         }
 
         Analytics.track(.multiSelectViewOverflowMenuRearrangeStarted, properties: ["source": actionDelegate.multiSelectViewSource])
@@ -243,16 +262,6 @@ class MultiSelectActionController: UIViewController, UITableViewDelegate, UITabl
         delegate.actionOrderChanged()
         dismiss(animated: true, completion: nil)
         Analytics.track(.multiSelectViewOverflowMenuRearrangeFinished, properties: ["source": actionDelegate.multiSelectViewSource])
-    }
-
-    private func enableScrollingIfViewTooSmall() {
-        guard actionsTable.isEditing else { return }
-
-        let newSize = view.systemLayoutSizeFitting(UIView.layoutFittingExpandedSize)
-        if newSize.height > UIScreen.main.bounds.height {
-            actionsTable.isScrollEnabled = true
-            actionsTable.bounces = false
-        }
     }
 
     func updateColors() {